project2-healthcare-lakehouse/requirements.txt
faker==24.0.0
pandas==2.2.0
pyarrow==15.0.0
pyspark==3.5.0
delta-spark==2.4.0
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.14.0
apache-airflow-providers-slack==8.5.0
great-expectations==0.18.8
boto3==1.34.0
tenacity==8.2.3
pytest==7.4.4
